backend/requirements.txt
fastapi
uvicorn[standard]>=0.27
taskiq-redis
supabase
python-dotenv
pydantic
yfinance>=0.2.36
requests>=2.31
beautifulsoup4>=4.12
transformers>=4.36
torch>=2.1
redis>=5.0
google-genai>=1.0
